refactor: changed files to lowercase
