+++ src/app/auth/login/page.tsx
@@ -1,5 +1,5 @@
 import { signInWithGithub } from './actions'
-import Login from '@/components/Auth/Login/Login'
+import Login from '@/components/auth/login/login'
 
 export default function LoginPage() {
   return (
